requirements_txt.txt
# MLflow 和模型训练
mlflow==2.9.2
scikit-learn==1.3.2

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 配置文件
pyyaml==6.0.1

# HTTP 请求（用于模型推理）
requests==2.31.0

# Dashboard 和可视化
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4

# 可选：Jupyter Notebook 支持
# jupyter==1.0.0
# notebook==7.0.6